added to Game
